Separate Meal Planner code from app.jsx
also updated my-react-app\package-lock.json version, resolved, and integrity.
